// src/app/pages/transfer/view-details-transfer/view-details-services/shared/view-details-services-enum.ts
export enum ViewDetailsServicesEnum {
    completeTitle = ' For Services',
    buttonSettings = 'Setup Service Settings',
    buttonDrivers = 'Setup Drivers',
    selectedTransferIdCookieName = 'selectedTransferId',

    plateTitle = 'Plate',
    childrenTitle = 'For Children',
    petsTitle = 'Allow Pets',
    doorTitle = 'Door to Door',
    stopTitle = 'Stops or Pickups',
    availabilityTitle = 'Availability',

    plate = 'plate',
    children = 'children',
    pets = 'pets',
    door = 'door',
    stop = 'stop',
    availability = 'availability',

    driversTitle = 'Drivers Asigned',
    bagsTitle = 'Bags Info',
    weight = 'Bags Weight',
    dimentions = 'Bags Dimentions',
    length = 'Length',
    width = 'Width',
    heigth = 'Height',
    schedule = 'Schedules',
    model = 'model',
    
    negativeValue = -1,
    space = ' ',
    kg = 'Kg',
    day = 'Day',
    night = 'Night',
}